chore: properly lazy load components and separate concerns to avoid unused javascript loaded
