+++ packages/webapp/components/footer/MobileFooterNavbar.tsx
@@ -5,17 +5,8 @@ import React, {
   useContext,
   useMemo,
 } from 'react';
-import { Flipper } from 'react-flip-toolkit';
 import classNames from 'classnames';
-import AuthContext from '@dailydotdev/shared/src/contexts/AuthContext';
-import { useRouter } from 'next/router';
-import ScrollToTopButton from '@dailydotdev/shared/src/components/ScrollToTopButton';
-import { Post } from '@dailydotdev/shared/src/graphql/posts';
-import { NewComment } from '@dailydotdev/shared/src/components/post/NewComment';
-import useActiveNav, {
-  UseActiveNav,
-} from '@dailydotdev/shared/src/hooks/useActiveNav';
-import { getFeedName } from '@dailydotdev/shared/src/lib/feed';
+import { Flipper } from 'react-flip-toolkit';
 import {
   AiIcon,
   BellIcon,
@@ -23,18 +14,19 @@ import {
   SourceIcon,
 } from '@dailydotdev/shared/src/components/icons';
 import { IconSize } from '@dailydotdev/shared/src/components/Icon';
+import useActiveNav, {
+  UseActiveNav,
+} from '@dailydotdev/shared/src/hooks/useActiveNav';
 import { squadCategoriesPaths } from '@dailydotdev/shared/src/lib/constants';
+import { useRouter } from 'next/router';
+import AuthContext from '@dailydotdev/shared/src/contexts/AuthContext';
+import { getFeedName } from '@dailydotdev/shared/src/lib/feed';
 import { useNotificationContext } from '@dailydotdev/shared/src/contexts/NotificationsContext';
 import { Bubble } from '@dailydotdev/shared/src/components/tooltips/utils';
 import { getUnreadText } from '@dailydotdev/shared/src/components/notifications/utils';
-import { FooterTab } from './footer/common';
-import { FooterNavBarTabs } from './footer/FooterNavBarTabs';
-import { FooterPlusButton } from './footer/FooterPlusButton';
-
-interface FooterNavBarProps {
-  showNav?: boolean;
-  post?: Post;
-}
+import { blurClasses, FooterTab } from './common';
+import { FooterPlusButton } from './FooterPlusButton';
+import { FooterNavBarTabs } from './FooterNavBarTabs';
 
 const Notifications = ({ active }: { active: boolean }): JSX.Element => {
   const { unreadCount } = useNotificationContext();
@@ -60,10 +52,11 @@ const selectedMapToTitle: Record<keyof UseActiveNav, string> = {
   profile: 'Profile',
 };
 
-export default function FooterNavBar({
-  showNav = false,
-  post,
-}: FooterNavBarProps): ReactElement {
+const MobileFooterNavbar = ({
+  isPostPage,
+}: {
+  isPostPage: boolean;
+}): ReactElement => {
   const router = useRouter();
   const { user, squads } = useContext(AuthContext);
   const feedName = getFeedName(router.pathname, { hasUser: !!user });
@@ -124,49 +117,25 @@ export default function FooterNavBar({
     return active?.title;
   }, [activeNav, router?.pathname, tabs]);
 
-  const blurClasses = 'bg-blur-baseline backdrop-blur-[2.5rem]';
   const activeClasses = classNames(
     blurClasses,
     'shadow-[0_4px_30px_rgba(0,0,0.1)]',
   );
 
   return (
-    <div
+    <Flipper
+      flipKey={activeTab}
+      spring="veryGentle"
+      element="nav"
       className={classNames(
-        'fixed !bottom-0 left-0 z-3 w-full',
-        showNav &&
-          'footer-navbar bg-gradient-to-t from-background-subtle from-70% to-transparent px-2 pt-2',
+        'grid w-full auto-cols-fr grid-flow-col items-center justify-between rounded-16',
+        !isPostPage && activeClasses,
+        !isPostPage && 'border-t border-border-subtlest-tertiary',
       )}
     >
-      {post ? (
-        <div className="my-2 w-full px-2 tablet:hidden">
-          <NewComment
-            post={post}
-            className={{
-              container: classNames(
-                blurClasses,
-                'h-12 shadow-[0_0.25rem_1.5rem_0_var(--theme-shadow-shadow1)]',
-              ),
-            }}
-          />
-        </div>
-      ) : (
-        <ScrollToTopButton />
-      )}
-      {showNav && (
-        <Flipper
-          flipKey={activeTab}
-          spring="veryGentle"
-          element="nav"
-          className={classNames(
-            'grid w-full auto-cols-fr grid-flow-col items-center justify-between rounded-16',
-            !post && activeClasses,
-            !post && 'border-t border-border-subtlest-tertiary',
-          )}
-        >
-          <FooterNavBarTabs activeTab={activeTab} tabs={tabs} />
-        </Flipper>
-      )}
-    </div>
+      <FooterNavBarTabs activeTab={activeTab} tabs={tabs} />
+    </Flipper>
   );
-}
+};
+
+export default MobileFooterNavbar;
